box annotation timestamp edit and type cleanup
